refactor: use snapshot string as Package.ID
