+++ pkg/dependency/parser/nodejs/pnpm/types.go
@@ -1,0 +1,44 @@
+package pnpm
+
+type PackageResolution struct {
+	Tarball string `yaml:"tarball,omitempty"`
+}
+
+type PackageInfo struct {
+	Resolution      PackageResolution `yaml:"resolution"`
+	Dependencies    map[string]string `yaml:"dependencies,omitempty"`
+	DevDependencies map[string]string `yaml:"devDependencies,omitempty"`
+	IsDev           bool              `yaml:"dev,omitempty"`
+	Name            string            `yaml:"name,omitempty"`
+	Version         string            `yaml:"version,omitempty"`
+}
+
+type LockFile struct {
+	LockfileVersion any                    `yaml:"lockfileVersion"`
+	Dependencies    map[string]any         `yaml:"dependencies,omitempty"`
+	DevDependencies map[string]any         `yaml:"devDependencies,omitempty"`
+	Packages        map[string]PackageInfo `yaml:"packages,omitempty"`
+
+	// V9
+	Importers map[string]Importer `yaml:"importers,omitempty"`
+	Snapshots map[string]Snapshot `yaml:"snapshots,omitempty"`
+}
+
+type Importer struct {
+	Dependencies    map[string]ImporterDepVersion `yaml:"dependencies,omitempty"`
+	DevDependencies map[string]ImporterDepVersion `yaml:"devDependencies,omitempty"`
+}
+
+type ImporterDepVersion struct {
+	Version string `yaml:"version,omitempty"`
+}
+
+type Snapshot struct {
+	Dependencies         map[string]string `yaml:"dependencies,omitempty"`
+	OptionalDependencies map[string]string `yaml:"optionalDependencies,omitempty"`
+}
+
+type snapshotPackage struct {
+	id        string
+	dependsOn []string
+}
